requirements.txt
asgiref==3.3.4
Django==3.2
django-environ==0.4.5
django-filter==2.4.0
django-leaflet==0.28.0
djangorestframework==3.12.4
djangorestframework-gis==0.17
psycopg2-binary==2.8.6
pytz==2021.1
sqlparse==0.4.1
uuid==1.30
